+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.14.0] - Unreleased
+## [1.14.0] - 2025-06-12
 ### Added
 - New `url_loader.ts` to load templates from URLs.
-- Support for destructuring in `for`. For example:
+- Support destructuring in `for`. For example:
   ```vto
   {{ for {name, age} of people }}
   {{ for index, {name, age} of people }}
@@ -357,7 +357,7 @@ First version
 [#106]: https://github.com/oscarotero/vento/issues/106
 [#108]: https://github.com/oscarotero/vento/issues/108
 
-[1.14.0]: https://github.com/oscarotero/vento/compare/v1.13.2...HEAD
+[1.14.0]: https://github.com/oscarotero/vento/compare/v1.13.2...v1.14.0
 [1.13.2]: https://github.com/oscarotero/vento/compare/v1.13.1...v1.13.2
 [1.13.1]: https://github.com/oscarotero/vento/compare/v1.13.0...v1.13.1
 [1.13.0]: https://github.com/oscarotero/vento/compare/v1.12.16...v1.13.0